requirements.txt
Flask
ultralytics
Pillow
torch
torchvision
numpy
